Added John Tomaselli's (@johntom) One-off charge
In the absence of a pull request, JohnTom requested the addition of and
provided the code for this machine which allows a charge to be created
(and subsequently captured with the capture charge machine) without
storing any card details. This request was made here:
https://github.com/particlebanana/machinepack-stripe/issues/1

I have made some minor cosmetic changes to the code provided by John to
keep the machine more in line with the rest of the pack.

Package.JSON updated to reflect this change, with the version bumped to
0.5.0 accordingly.